Moved DataElement.from_json helper methods to own class
- moved empty_value_for_VR to config to avoid circular dependencies
